Port some more gnarly definitions
